Removed the massive code duplication in the 'outdated' command
