+++ commands/instances.go
@@ -522,124 +522,6 @@ func UpdateCoreLibrariesIndex(ctx context.Context, req *rpc.UpdateCoreLibrariesI
 	return nil
 }
 
-// Outdated returns a list struct containing both Core and Libraries that can be updated
-func Outdated(ctx context.Context, req *rpc.OutdatedRequest) (*rpc.OutdatedResponse, error) {
-	id := req.GetInstance().GetId()
-
-	lm := GetLibraryManager(id)
-	if lm == nil {
-		return nil, &arduino.InvalidInstanceError{}
-	}
-
-	outdatedLibraries := []*rpc.InstalledLibrary{}
-	for _, libAlternatives := range lm.Libraries {
-		for _, library := range libAlternatives.Alternatives {
-			if library.Location != libraries.User {
-				continue
-			}
-			available := lm.Index.FindLibraryUpdate(library)
-			if available == nil {
-				continue
-			}
-
-			outdatedLibraries = append(outdatedLibraries, &rpc.InstalledLibrary{
-				Library: getOutputLibrary(library),
-				Release: getOutputRelease(available),
-			})
-		}
-	}
-
-	pm := GetPackageManager(id)
-	if pm == nil {
-		return nil, &arduino.InvalidInstanceError{}
-	}
-
-	outdatedPlatforms := []*rpc.Platform{}
-	for _, targetPackage := range pm.Packages {
-		for _, installed := range targetPackage.Platforms {
-			if installedRelease := pm.GetInstalledPlatformRelease(installed); installedRelease != nil {
-				latest := installed.GetLatestRelease()
-				if latest == nil || latest == installedRelease {
-					continue
-				}
-				rpcPlatform := PlatformReleaseToRPC(latest)
-				rpcPlatform.Installed = installedRelease.Version.String()
-
-				outdatedPlatforms = append(
-					outdatedPlatforms,
-					rpcPlatform,
-				)
-			}
-		}
-	}
-
-	return &rpc.OutdatedResponse{
-		OutdatedLibraries: outdatedLibraries,
-		OutdatedPlatforms: outdatedPlatforms,
-	}, nil
-}
-
-func getOutputLibrary(lib *libraries.Library) *rpc.Library {
-	insdir := ""
-	if lib.InstallDir != nil {
-		insdir = lib.InstallDir.String()
-	}
-	srcdir := ""
-	if lib.SourceDir != nil {
-		srcdir = lib.SourceDir.String()
-	}
-	utldir := ""
-	if lib.UtilityDir != nil {
-		utldir = lib.UtilityDir.String()
-	}
-	cntplat := ""
-	if lib.ContainerPlatform != nil {
-		cntplat = lib.ContainerPlatform.String()
-	}
-
-	return &rpc.Library{
-		Name:              lib.Name,
-		Author:            lib.Author,
-		Maintainer:        lib.Maintainer,
-		Sentence:          lib.Sentence,
-		Paragraph:         lib.Paragraph,
-		Website:           lib.Website,
-		Category:          lib.Category,
-		Architectures:     lib.Architectures,
-		Types:             lib.Types,
-		InstallDir:        insdir,
-		SourceDir:         srcdir,
-		UtilityDir:        utldir,
-		Location:          lib.Location.ToRPCLibraryLocation(),
-		ContainerPlatform: cntplat,
-		Layout:            lib.Layout.ToRPCLibraryLayout(),
-		RealName:          lib.RealName,
-		DotALinkage:       lib.DotALinkage,
-		Precompiled:       lib.Precompiled,
-		LdFlags:           lib.LDflags,
-		IsLegacy:          lib.IsLegacy,
-		Version:           lib.Version.String(),
-		License:           lib.License,
-	}
-}
-
-func getOutputRelease(lib *librariesindex.Release) *rpc.LibraryRelease {
-	if lib != nil {
-		return &rpc.LibraryRelease{
-			Author:        lib.Author,
-			Version:       lib.Version.String(),
-			Maintainer:    lib.Maintainer,
-			Sentence:      lib.Sentence,
-			Paragraph:     lib.Paragraph,
-			Website:       lib.Website,
-			Category:      lib.Category,
-			Architectures: lib.Architectures,
-			Types:         lib.Types,
-		}
-	}
-	return &rpc.LibraryRelease{}
-}
-
 // LoadSketch collects and returns all files composing a sketch
 func LoadSketch(ctx context.Context, req *rpc.LoadSketchRequest) (*rpc.LoadSketchResponse, error) {
 	// TODO: This should be a ToRpc function for the Sketch struct
