+++ commands/outdated/outdated.go
@@ -1,0 +1,48 @@
+// This file is part of arduino-cli.
+//
+// Copyright 2022 ARDUINO SA (http://www.arduino.cc/)
+//
+// This software is released under the GNU General Public License version 3,
+// which covers the main part of arduino-cli.
+// The terms of this license can be found at:
+// https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// You can be released from the requirements of the above licenses by purchasing
+// a commercial license. Buying such a license is mandatory if you want to
+// modify or otherwise use the software for commercial activities involving the
+// Arduino software without disclosing the source code of your own applications.
+// To purchase a commercial license, send an email to license@arduino.cc.
+
+package outdated
+
+import (
+	"context"
+
+	"github.com/arduino/arduino-cli/commands/core"
+	"github.com/arduino/arduino-cli/commands/lib"
+	rpc "github.com/arduino/arduino-cli/rpc/cc/arduino/cli/commands/v1"
+)
+
+// Outdated returns a list struct containing both Core and Libraries that can be updated
+func Outdated(ctx context.Context, req *rpc.OutdatedRequest) (*rpc.OutdatedResponse, error) {
+	libraryListResponse, err := lib.LibraryList(ctx, &rpc.LibraryListRequest{
+		Instance:  req.GetInstance(),
+		Updatable: true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	getPlatformsResp, err := core.GetPlatforms(&rpc.PlatformListRequest{
+		Instance:      req.GetInstance(),
+		UpdatableOnly: true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return &rpc.OutdatedResponse{
+		OutdatedLibraries: libraryListResponse.GetInstalledLibraries(),
+		OutdatedPlatforms: getPlatformsResp,
+	}, nil
+}
